добавлены цвет и модель

+++ src/main/java/abc/vaadin/data/entity/Cart.java
@@ -4,30 +4,30 @@ import jakarta.persistence.Entity;
 
 @Entity
 public class Cart extends AbstractEntity{
-    Long product_id;
-    Long user_id;
+    Integer product_id;
+    Integer user_id;
 
     public Cart() {
     }
 
-    public Cart(Long product_id, Long user_id) {
+    public Cart(Integer product_id, Integer user_id) {
         this.product_id = product_id;
         this.user_id = user_id;
     }
 
-    public Long getProduct_id() {
+    public Integer getProduct_id() {
         return product_id;
     }
 
-    public void setProduct_id(Long product_id) {
+    public void setProduct_id(Integer product_id) {
         this.product_id = product_id;
     }
 
-    public Long getUser_id() {
+    public Integer getUser_id() {
         return user_id;
     }
 
-    public void setUser_id(Long user_id) {
+    public void setUser_id(Integer user_id) {
         this.user_id = user_id;
     }
 }
